+++ source/audio/midioutputdevice.cpp
@@ -67,12 +67,16 @@ MidiOutputDevice::~MidiOutputDevice()
 {
     // Make sure there aren't any lingering notes.
     if (myMidiOut)
+        stopAllNotes();
+}
+
+void
+MidiOutputDevice::stopAllNotes()
+{
+    for (uint8_t channel = 0; channel < Midi::NUM_MIDI_CHANNELS_PER_PORT;
+         ++channel)
     {
-        for (uint8_t channel = 0; channel < Midi::NUM_MIDI_CHANNELS_PER_PORT;
-             ++channel)
-        {
-            sendMidiMessage(ControlChange + channel, AllNotesOff, 0);
-        }
+        sendMidiMessage(ControlChange + channel, AllNotesOff, 0);
     }
 }